BackFace Culling: take account of Element "closeness"
In Element class, a parameter has been added to define if the Element is
closed or not (open). E.g. a Sphere is closed but a Cylinder is open
(default for an Element) unless top and bottom are added (which is not
the case in current implementation).
This parameter is taken into account at rendering and combined with the
RenderContext.backaceCulling parameter to eliminate or not the back
faces (triangles) pointing in opposite direction than viewer.

Next stage will be to create "closed" versions of Cone, Cylinder and
Pyramid Elements (currently open since no top nor bottom are created for
these Elements in current implementation).