Fix bug in Point light calculation: vector of light was opposite.
Still the convention used for light direction (a vector pointing to
light) that was used since the beginning for Directional Light seems
weird. It would be preferable to use the direction of the light itself
(and to change the test of dotNL sign in Rasterizer appropriately.
It may however have an impact on all tests using Directional Light (need
to change light vector direction).
-> Future change.