DirectionalLight direction vector now direction FROM light.
DirectionalLight direction vector is now the direction FROM light for a
more obvious usage of the light direction: this is the direction of
light rays. While light vector remains the calculated vector to light
from a surface point (so it points to the light) to calculate the dotNL
product.
This is a significant behavior change that impacts most of the Test
classes.

Some other clean-up.