PerspectiveContext improvements + FPS counter (Stats traces)
Improve Perspective context to calculate a perspective and defining a
pixel width for the map. This is useful for the ShadowMap calculation to
keep a constant pixel size while having a variable width.
Use Bounding Box dimension to calculate the PerspectiveContext.
Identify BUG : World Bounding Box does not take account of Element's
transformations (translations etc.).

Implement FPS counter in RenderEngine. For the moment it is dieplayed as
Stats traces (need enablement of this level).